# openapi/bidvector-operator-api.yaml
openapi: 3.0.3
info:
  title: bid-vector operator API
  version: "6a2b"
  description: >-
    단일 운영자 소비자 전용(운영자 결정 2026-09-16 ①). 이 문서가 단일 출처다(D-6A1-8,
    수작성) — `OpenApiContractTest`가 구현과 대조한다. M6/6A-1이 전략 조회(읽기 전용)를,
    M6/6A-3+6F-3(D-6A3-11)이 평가 dry-run endpoint를 연다 — 둘 다 커밋을 만들지 않는다
    (실행·명령 경로 없음). M6/6A-2b(D-6A2b-1)가 편집 세션 endpoint 여섯을 열어 전략을
    **쓰는** 길을 낸다 — 모든 쓰기는 편집 세션을 지나고 한 트랜잭션에서 커밋된다.

    **405 와 406 은 어느 operation 아래에도 적지 않는다.** 405 는 operation 의 응답이 아니라
    「그 경로에 그 메서드가 없다」는 사실이라 적을 자리가 없다. 406 은 operation 의 응답이
    맞지만 **본문이 없다**(D-6A2b-29) — 클라이언트가 JSON 을 받지 않겠다고 한 요청이라
    `ErrorBody` 를 쓸 수 없고, 그래서 `ErrorBody.code` 어휘에도 그 코드를 두지 않는다.
    구현은 매핑된 모든 경로에서 미선언 메서드에 405 `ErrorBody` 를, 받아들일 수 없는 `Accept`
    에 본문 없는 406 을 낸다 — `ProductionHttpSurfaceTest` 가 출하 조립의 등록된 매핑 전수로
    두 축을 함께 잰다.
servers:
  - url: /
security:
  - operatorCredential: []
paths:
  /api/strategy:
    get:
      operationId: readStrategy
      summary: 현재 운영자 전략 조회
      responses:
        "200":
          description: 현재 전략
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyReadResponse"
        "401":
          description: >-
            자격증명 없음 또는 틀림 — 사유를 구분하지 않는다(D-6A1-6).
            code: UNAUTHENTICATED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          description: >-
            내부 오류(저장된 전략이 유효하지 않음 포함) 또는 감사 기록 실패(D-6A1-17,
            fail-closed). code: INVALID_STORED_STRATEGY | INTERNAL_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/evaluation-dry-runs:
    post:
      operationId: evaluateDryRun
      summary: 평가 dry-run(D-6A3-1·6 — 외부 effect 없음, outbox 미접촉)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluationDryRunRequest"
      responses:
        "200":
          description: 판정 분포(결과별 공고 ID 배열 + 건수)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationDryRunResponse"
        "400":
          description: >-
            currentActiveBids 가 유효한 JSON 정수(0..2147483647)가 아니다(D-6A3-5·D-6A3-19) —
            음수, 누락, null, 문자열·소수(조용한 강제 변환 없음), 범위 초과, 본문이 JSON
            object 가 아님, 비JSON, 빈 본문을 모두 포함한다. code: INVALID_REQUEST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: "자격증명 없음 또는 틀림(D-6A1-6). code: UNAUTHENTICATED"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: >-
            전략에 여력 상한 미설정(D-6A3-4, code: MAX_ACTIVE_BIDS_NOT_CONFIGURED) 또는
            후보 스캔이 상한 초과(D-6A3-7, code: CANDIDATE_CAP_EXCEEDED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "415":
          description: >-
            요청 본문이 application/json 이 아니다(M6/6A-2b D-6A2b-7).
            code: UNSUPPORTED_MEDIA_TYPE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          description: >-
            내부 오류(저장된 전략이 유효하지 않음 포함) 또는 감사 기록 실패(D-6A1-17,
            fail-closed). code: INVALID_STORED_STRATEGY | INTERNAL_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/strategy/edit-sessions:
    post:
      operationId: beginEditSession
      summary: 편집 세션 시작(D-6A2b-1) — 세션 id 는 서버가 만든다
      description: >-
        409 `SESSION_ALREADY_ACTIVE` 는 선언하지 않는다(C-3) — 세션 id 를 서버가 UUID 로
        만들므로 「같은 id 에 이미 비종단 세션」 가드에 HTTP 로 닿을 길이 없다. 그 코드는
        use case 의 전수 `when` 을 총(total)으로 유지하려고 남아 있을 뿐이고, 도달 불가는
        알려진 제한으로 등재돼 있다. 버려진 세션의 처분은 `OPEN-6A2B-ABANDONED-SESSIONS`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditSessionBeginRequest"
      responses:
        "201":
          description: 새 세션(state=WAITING_FOR_VALUE)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditSessionResponse"
        "400":
          description: "본문 형식 오류·알 수 없는 field 토큰. code: INVALID_REQUEST"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: "자격증명 없음 또는 틀림(D-6A1-6). code: UNAUTHENTICATED"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "415":
          description: "요청 본문이 application/json 이 아니다. code: UNSUPPORTED_MEDIA_TYPE"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          description: >-
            내부 오류(저장된 전략이 유효하지 않음 포함) 또는 감사 기록 실패(D-6A1-17).
            code: INVALID_STORED_STRATEGY | INTERNAL_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/strategy/edit-sessions/{sessionId}:
    get:
      operationId: readEditSession
      summary: 편집 세션 조회 — 접근 시점에 만료를 먼저 접는다(D-6A2b-11)
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 세션 상태
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditSessionResponse"
        "401":
          description: "자격증명 없음 또는 틀림. code: UNAUTHENTICATED"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "400":
          description: >-
            경로 변수가 비어 있다(`%20` 등 — 디코딩 뒤 공백). code: INVALID_REQUEST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: "그 id 의 세션이 없다. code: SESSION_NOT_FOUND"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: >-
            조회가 접는 만료 fold 의 영속이 동시 전진과 부딪쳤다(C-2).
            code: EDIT_SESSION_CONFLICT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          description: "내부 오류. code: INTERNAL_ERROR"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/strategy/edit-sessions/{sessionId}/value:
    post:
      operationId: provideEditSessionValue
      summary: 값 제출(D-6A2b-2) — 현재 전략에서 그 필드 하나만 바꾼 초안을 만든다
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditSessionValueRequest"
      responses:
        "200":
          description: 세션 상태(값이 유효하면 WAITING_FOR_CONFIRMATION)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditSessionResponse"
        "400":
          description: >-
            본문 형식 오류(code: INVALID_REQUEST) 또는 전략 값 불변식 위반
            (code: STRATEGY_VALUE_INVALID — 어느 불변식인지는 싣지 않는다,
            OPEN-6A2B-VIOLATION-DETAIL)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: "자격증명 없음 또는 틀림. code: UNAUTHENTICATED"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: "그 id 의 세션이 없다. code: SESSION_NOT_FOUND"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: >-
            세션이 이 명령을 받을 수 없다. code: SESSION_EXPIRED | IDEMPOTENCY_CONFLICT |
            ACTOR_MISMATCH | SYSTEM_ACTOR_NOT_PERMITTED | STALE_REVISION |
            INVALID_TRANSITION | SESSION_ALREADY_ACTIVE | EDIT_SESSION_CONFLICT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "415":
          description: "요청 본문이 application/json 이 아니다. code: UNSUPPORTED_MEDIA_TYPE"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          description: "내부 오류. code: INVALID_STORED_STRATEGY | INTERNAL_ERROR"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/strategy/edit-sessions/{sessionId}/confirm:
    post:
      operationId: confirmEditSession
      summary: 확인 — seenRevision 이 지금 전략과 다르면 거부한다(우회 3)
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditSessionConfirmRequest"
      responses:
        "200":
          description: 세션 상태(적용되면 APPLIED · strategyRevision 이 1 오른다)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditSessionResponse"
        "400":
          description: "본문 형식 오류. code: INVALID_REQUEST"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: "자격증명 없음 또는 틀림. code: UNAUTHENTICATED"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: "그 id 의 세션이 없다. code: SESSION_NOT_FOUND"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: >-
            세션이 이 명령을 받을 수 없다(사유 코드는 value 와 같은 목록).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "415":
          description: "요청 본문이 application/json 이 아니다. code: UNSUPPORTED_MEDIA_TYPE"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          description: "내부 오류. code: INVALID_STORED_STRATEGY | INTERNAL_ERROR"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/strategy/edit-sessions/{sessionId}/edit:
    post:
      operationId: requestEditSessionEdit
      summary: 확인 화면에서 (같거나 다른) 필드로 되돌아간다
      description: >-
        **확인 대기(WAITING_FOR_CONFIRMATION) 에서만 유효하다**(L-r2-4). 값 대기 상태에서
        필드를 바꿀 길은 없다 — `POST /{sessionId}/value` 는 세션이 기다리는 필드만 받고
        (D-6A2b-25), 되돌아가기는 확인 대기에서만 전이표에 있다. 잘못 연 세션은 취소하고
        다시 연다. 그 밖의 상태에서는 409 `INVALID_TRANSITION` 이다.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditSessionEditRequest"
      responses:
        "200":
          description: 세션 상태(WAITING_FOR_VALUE)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditSessionResponse"
        "400":
          description: "본문 형식 오류·알 수 없는 field 토큰. code: INVALID_REQUEST"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: "자격증명 없음 또는 틀림. code: UNAUTHENTICATED"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: "그 id 의 세션이 없다. code: SESSION_NOT_FOUND"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: "세션이 이 명령을 받을 수 없다(사유 코드는 value 와 같은 목록)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "415":
          description: "요청 본문이 application/json 이 아니다. code: UNSUPPORTED_MEDIA_TYPE"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          description: "내부 오류. code: INVALID_STORED_STRATEGY | INTERNAL_ERROR"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/strategy/edit-sessions/{sessionId}/cancel:
    post:
      operationId: cancelEditSession
      summary: 취소 — 종단 상태다(재개는 새 세션, D-4A-4)
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditSessionCancelRequest"
      responses:
        "200":
          description: 세션 상태(CANCELLED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditSessionResponse"
        "400":
          description: "본문 형식 오류. code: INVALID_REQUEST"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: "자격증명 없음 또는 틀림. code: UNAUTHENTICATED"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: "그 id 의 세션이 없다. code: SESSION_NOT_FOUND"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: "세션이 이 명령을 받을 수 없다(사유 코드는 value 와 같은 목록)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "415":
          description: "요청 본문이 application/json 이 아니다. code: UNSUPPORTED_MEDIA_TYPE"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          description: "내부 오류. code: INVALID_STORED_STRATEGY | INTERNAL_ERROR"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /{unmatched}:
    get:
      operationId: unmatchedPath
      summary: 미매핑 경로 — 계약 문서화 목적(D-6A1-21, 실제로는 어떤 controller도 이 경로를 다루지 않는다)
      responses:
        "404":
          description: "등록된 endpoint가 없음. code: NOT_FOUND"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
components:
  securitySchemes:
    operatorCredential:
      type: apiKey
      in: header
      name: X-Operator-Credential
  schemas:
    # D-6A1-20 ⓑ — 평탄하다는 것을 이 스키마 자신이 단언한다: 모든 속성이 string·integer·
    # number·array<string> 중 하나이고 중첩 object 속성이 없다(`OpenApiContractTest`가
    # 이 문서에서 object 타입 속성이 새로 생기면 실패하도록 재확인한다).
    StrategyReadResponse:
      type: object
      required:
        - revision
        - focusCategories
        - focusRegionTerms
        - excludeRegionTerms
        - requiredKeywordTerms
        - excludeKeywordTerms
      properties:
        revision:
          type: integer
        focusCategories:
          type: array
          items:
            type: string
        focusRegionTerms:
          type: array
          items:
            type: string
        excludeRegionTerms:
          type: array
          items:
            type: string
        requiredKeywordTerms:
          type: array
          items:
            type: string
        excludeKeywordTerms:
          type: array
          items:
            type: string
        minBudgetWon:
          type: integer
          format: int64
          nullable: true
        minBudgetCurrency:
          type: string
          nullable: true
        minBudgetVatTreatment:
          type: string
          nullable: true
        minBudgetProvenance:
          type: string
          nullable: true
        maxBudgetWon:
          type: integer
          format: int64
          nullable: true
        maxBudgetCurrency:
          type: string
          nullable: true
        maxBudgetVatTreatment:
          type: string
          nullable: true
        maxBudgetProvenance:
          type: string
          nullable: true
        minimumMatchScore:
          type: number
          nullable: true
        minimumProbabilityScore:
          type: number
          nullable: true
        bidNowThreshold:
          type: number
          nullable: true
        reviewThreshold:
          type: number
          nullable: true
        candidateLimit:
          type: integer
          nullable: true
        maxActiveBids:
          type: integer
          nullable: true
    EvaluationDryRunRequest:
      type: object
      required:
        - currentActiveBids
      properties:
        currentActiveBids:
          type: integer
          format: int32
          minimum: 0
    # D-6A3-6 — 평탄하다. 결과별 공고 ID 배열 + 건수만 싣는다(후보별 사유 상세는
    # OPEN-6A3-EVALUATION-DETAIL). 공고명·기관명 원문은 없다(privacy).
    EvaluationDryRunResponse:
      type: object
      required:
        - strategyRevision
        - candidateCount
        - currentActiveBids
        - maxActiveBids
        - bidNowNoticeIds
        - reviewNoticeIds
        - skipNoticeIds
        - notReachedNoticeIds
        - wouldNotifyNoticeIds
      properties:
        strategyRevision:
          type: integer
        candidateCount:
          type: integer
        currentActiveBids:
          type: integer
        maxActiveBids:
          type: integer
        bidNowNoticeIds:
          type: array
          items:
            type: string
        reviewNoticeIds:
          type: array
          items:
            type: string
        skipNoticeIds:
          type: array
          items:
            type: string
        notReachedNoticeIds:
          type: array
          items:
            type: string
        wouldNotifyNoticeIds:
          type: array
          items:
            type: string
    # D-6A3-20(검토 라운드 1 contract-keeper V2) — code 는 ErrorMapping.forThrowable 이 낼 수
    # 있는 전 코드의 열거다. `OpenApiContractTest`가 app.http.ErrorCode 의 상수 집합과 이
    # 목록을 양방향(부분집합 아니라 등식)으로 대조한다 — 한쪽에만 코드를 더하면 RED.
    # `type: string` + `enum` 은 D-6A1-38 평탄 술어의 NON_FLAT_KEYS 밖이라 게이트는 그대로
    # 초록이다.
    # M6/6A-2b D-6A2b-1 — 편집 요청·응답 다섯. 전부 평탄하다(D-6A1-20 ⓑ). `field` 의 enum 은
    # 구현의 필드 어휘 전수와 **집합 등식**으로 대조된다(`OpenApiEditSessionContractTest`) —
    # 새 필드를 코드에만 더하거나 문서에만 더하면 RED 다.
    EditSessionBeginRequest:
      type: object
      required:
        - field
      properties:
        field:
          type: string
          enum:
            - FOCUS_CATEGORY
            - FOCUS_REGION
            - EXCLUDE_REGION
            - REQUIRED_KEYWORD
            - EXCLUDE_KEYWORD
            - MIN_BUDGET
            - MAX_BUDGET
            - MINIMUM_MATCH_SCORE
            - MINIMUM_PROBABILITY_SCORE
            - BID_NOW_THRESHOLD
            - REVIEW_THRESHOLD
            - CANDIDATE_LIMIT
            - MAX_ACTIVE_BIDS
    # 값 칸 넷 가운데 `field` 가 이름하는 하나만 싣는다 — 다른 칸이 함께 오면 400 이다
    # (조용히 무시하지 않는다). 칸 배정: 업종·지역·키워드 다섯 → terms, 예산 둘 → amountWon,
    # 임계 넷 → number, 후보 상한·여력 상한 → count.
    # L-3 — 「`field` 가 이름하는 칸 **하나**가 필수」는 이 스키마 형태로 표현할 수 없다
    # (`oneOf` 는 평탄 게이트에 걸린다). 구현은 엄격하게 강제한다: 그 칸이 없으면 400 이고,
    # 다른 칸이 함께 오거나 알려진 키 밖이 오면 역시 400 이다. 칸 배정은 위 `field` enum 순서와
    # 같다 — 업종·지역·키워드 다섯은 `terms`, 예산 둘은 `amountWon`, 임계 넷은 `number`,
    # 후보 상한·여력 상한은 `count`.
    EditSessionValueRequest:
      type: object
      required:
        - commandId
        - field
      properties:
        commandId:
          type: string
        field:
          type: string
          enum:
            - FOCUS_CATEGORY
            - FOCUS_REGION
            - EXCLUDE_REGION
            - REQUIRED_KEYWORD
            - EXCLUDE_KEYWORD
            - MIN_BUDGET
            - MAX_BUDGET
            - MINIMUM_MATCH_SCORE
            - MINIMUM_PROBABILITY_SCORE
            - BID_NOW_THRESHOLD
            - REVIEW_THRESHOLD
            - CANDIDATE_LIMIT
            - MAX_ACTIVE_BIDS
        terms:
          type: array
          items:
            type: string
        amountWon:
          type: integer
          format: int64
          minimum: 0
        number:
          type: number
        count:
          type: integer
          format: int32
    EditSessionConfirmRequest:
      type: object
      required:
        - commandId
        - seenRevision
      properties:
        commandId:
          type: string
        seenRevision:
          type: integer
          format: int32
          minimum: 0
    EditSessionEditRequest:
      type: object
      required:
        - commandId
        - field
      properties:
        commandId:
          type: string
        field:
          type: string
          enum:
            - FOCUS_CATEGORY
            - FOCUS_REGION
            - EXCLUDE_REGION
            - REQUIRED_KEYWORD
            - EXCLUDE_KEYWORD
            - MIN_BUDGET
            - MAX_BUDGET
            - MINIMUM_MATCH_SCORE
            - MINIMUM_PROBABILITY_SCORE
            - BID_NOW_THRESHOLD
            - REVIEW_THRESHOLD
            - CANDIDATE_LIMIT
            - MAX_ACTIVE_BIDS
    EditSessionCancelRequest:
      type: object
      required:
        - commandId
      properties:
        commandId:
          type: string
    # 초안(draft) 전체는 싣지 않는다 — 확인 중인 값은 방금 보낸 값이고 현재 전략은
    # `GET /api/strategy` 가 낸다.
    EditSessionResponse:
      type: object
      required:
        - sessionId
        - state
        - sessionVersion
        - expiresAt
      properties:
        sessionId:
          type: string
        state:
          type: string
          enum:
            - WAITING_FOR_VALUE
            - WAITING_FOR_CONFIRMATION
            - APPLIED
            - CANCELLED
            - EXPIRED
        field:
          type: string
          nullable: true
          # C-4 — 종단 상태에서 실제로 `null` 이 나간다. 엄격한 검증기가 그 값을 거부하지
          # 않도록 어휘에 함께 싣는다(대조 test 는 `null` 을 뺀 집합으로 등식을 잰다).
          enum:
            - null
            - FOCUS_CATEGORY
            - FOCUS_REGION
            - EXCLUDE_REGION
            - REQUIRED_KEYWORD
            - EXCLUDE_KEYWORD
            - MIN_BUDGET
            - MAX_BUDGET
            - MINIMUM_MATCH_SCORE
            - MINIMUM_PROBABILITY_SCORE
            - BID_NOW_THRESHOLD
            - REVIEW_THRESHOLD
            - CANDIDATE_LIMIT
            - MAX_ACTIVE_BIDS
        sessionVersion:
          type: integer
        expiresAt:
          type: string
          format: date-time
        strategyRevision:
          type: integer
          nullable: true
        baseRevision:
          type: integer
          nullable: true
    ErrorBody:
      type: object
      required:
        - code
        - message
        - correlationId
      properties:
        code:
          type: string
          enum:
            - UNAUTHENTICATED
            - NOT_FOUND
            - INVALID_STORED_STRATEGY
            - INTERNAL_ERROR
            - MAX_ACTIVE_BIDS_NOT_CONFIGURED
            - CANDIDATE_CAP_EXCEEDED
            - INVALID_REQUEST
            - SESSION_NOT_FOUND
            - SESSION_EXPIRED
            - IDEMPOTENCY_CONFLICT
            - ACTOR_MISMATCH
            - SYSTEM_ACTOR_NOT_PERMITTED
            - STALE_REVISION
            - INVALID_TRANSITION
            - SESSION_ALREADY_ACTIVE
            - EDIT_SESSION_CONFLICT
            - STRATEGY_VALUE_INVALID
            - METHOD_NOT_ALLOWED
            - UNSUPPORTED_MEDIA_TYPE
        message:
          type: string
        correlationId:
          type: string